Standardize type usage, improve resource handling, and clean up codebase
- Refactored `ResultSetTableMaxColumnWidth` and `ResultSetTableMaxRows` to use `int` instead of `long`.
- Switched to try-with-resources for loading properties files to ensure proper resource management.
- Simplified `PreparedStatement` argument tracing by removing unnecessary type hints.
- Updated logging header formats and optimized boolean formatting in `RdbmsSpecifics`.
- Replaced `StringBuffer` with more efficient `StringBuilder` where applicable.
- Made selected fields final for thread safety improvements and code clarity.

+++ src/main/java/net/sf/log4jdbc/sql/resultsetcollector/ResultSetCollectorPrinter.java
@@ -61,7 +61,7 @@ public class ResultSetCollectorPrinter {
             numeric[column - 1] = isNumericType(resultSetCollector.getColumnType(column));
         }
 
-        int maxColumnWidth = (int) Properties.getResultSetTableMaxColumnWidth();
+        int maxColumnWidth = Properties.getResultSetTableMaxColumnWidth();
 
         // 2. Calculate maximum width of each column (including column name)
         for (int column = 1; column <= columnCount; column++) {
@@ -110,7 +110,7 @@ public class ResultSetCollectorPrinter {
         table.append(createBorderLine(maxLength, '├', '┼', '┤'));
 
         // 8. Data rows
-        int maxRows = (int) Properties.getResultSetTableMaxRows();
+        int maxRows = Properties.getResultSetTableMaxRows();
         if (resultSetCollector.getRows() != null) {
             int rowIndex = 0;
             for (List<Object> printRow : resultSetCollector.getRows()) {
